backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
fastapi-mail
python-dotenv
gunicorn